separate metrics to different files/dependency groups
